post submit  tags
Finally, got to submit all the checkboxes related to the tags in the edit post area.

I create checkbox inputs dinamically for every tag in the post, and now, updating or saving the post uses the http post protocol wordpress api to save the data.

Later, I'll use that data to add references to the post content area