added link in tag archives title
Changed functions to admin class, had not sense for them to be at public class of the plugin.

It would be interesting to study add_post_meta, but will only work with posts, not tag pages.

added option to change tag archive title for a link to wikidata, just in case the web owner/admin wants to